+++ packages/shared/pkg/atomicbitset/bitset.go
@@ -18,18 +18,18 @@ func New() *Bitset {
 	}
 }
 
-func (b *Bitset) HasRange(start, end int64) bool {
+func (b *Bitset) HasRange(start, end uint64) bool {
 	b.mu.RLock()
 	defer b.mu.RUnlock()
 
-	return b.bm.CardinalityInRange(uint64(start), uint64(end)) == uint64(end-start)
+	return b.bm.CardinalityInRange(start, end) == end-start
 }
 
-func (b *Bitset) SetRange(start, end int64) {
+func (b *Bitset) SetRange(start, end uint64) {
 	b.mu.Lock()
 	defer b.mu.Unlock()
 
-	b.bm.AddRange(uint64(start), uint64(end))
+	b.bm.AddRange(start, end)
 }
 
 func (b *Bitset) BitSet() *bitset.BitSet {
